Fixed highscore
