+++ src/components/HighScore/index.tsx
@@ -1,32 +1,8 @@
-import { useEffect, useState } from 'react';
-import { IState } from '../../App';
-export default function HighScore({ state }: { state: IState }) {
-  const { board, playerRed, playerYellow } = state;
+export default function HighScore() {
   const highscoresData = JSON.parse(localStorage.getItem('highscores') || '[]') as { name: string; moves: number }[];
   const sortedHighscores = highscoresData.sort((a, b) => a.moves - b.moves).slice(0, 10);
-  const highScore = (name: string, moves: number) => {
-    const highscores = JSON.parse(localStorage.getItem('highscores') || '[]') as { name: string; moves: number }[];
-
-    highscores.push({ name, moves });
-    highscores.sort((a, b) => a.moves - b.moves).slice(0, 10);
-
-    localStorage.setItem('highscores', JSON.stringify(highscores));
-  };
-
-  const [scoreUpdated, setScoreUpdated] = useState(false);
-
-  useEffect(() => {
-    if (board.gameOver && board.winner && !scoreUpdated) {
-      const winnerName = state.board.winner === 'Red' ? playerRed!.name : playerYellow!.name;
-
-      const winnerMoves = state.board.winner === 'Yellow' ? board.moveCounterRed : board.moveCounterYellow;
-      highScore(winnerName, winnerMoves);
-
-      setScoreUpdated(true);
-    }
-  }, [board.gameOver, board.winner, scoreUpdated]);
   return (
-    <div className='highscore-list'>
+    <div className="highscore-list">
       <h3>Highscore List</h3>
       <table>
         <thead>
@@ -39,7 +15,7 @@ export default function HighScore({ state }: { state: IState }) {
           {sortedHighscores.map((list, index) => (
             <tr key={index}>
               <td>{list.name}</td>
-              <td className=' highscore-moves'>{list.moves}</td>
+              <td className=" highscore-moves">{list.moves}</td>
             </tr>
           ))}
         </tbody>
